+++ libs/shared-module/src/lib/components/translation-item-simple/translation-item-simple.component.ts
@@ -1,22 +1,21 @@
 import { Component, Inject, Input, OnChanges, OnDestroy } from '@angular/core';
 import {
   ControlContainer,
-  FormBuilder,
   FormControl,
   FormGroupDirective,
+  UntypedFormBuilder,
   UntypedFormControl,
 } from '@angular/forms';
 import { IXliffInterpolation, IXliffPlural } from '@vtabary/xliff2js';
 
-// TODO duplication de TranslationItemComponent a supprimer
 @Component({
-  selector: 'translatol-translation-plural-item',
-  templateUrl: './translation-plural-item.component.html',
-  styleUrls: ['./translation-plural-item.component.scss'],
+  selector: 'translatol-translation-item-simple',
+  templateUrl: './translation-item-simple.component.html',
+  styleUrls: ['./translation-item-simple.component.scss'],
 })
-export class TranslationPluralItemComponent implements OnChanges, OnDestroy {
+export class TranslationItemSimpleComponent implements OnChanges, OnDestroy {
   @Input()
-  public source?: string | IXliffInterpolation | IXliffPlural;
+  public source?: string;
 
   @Input()
   public target?: string | IXliffInterpolation | IXliffPlural;
@@ -31,30 +30,14 @@ export class TranslationPluralItemComponent implements OnChanges, OnDestroy {
    * @internal
    */
   public control?: UntypedFormControl;
-  /**
-   * @internal
-   */
-  public text?: string;
-  /**
-   * @internal
-   */
-  public interpolation?: IXliffInterpolation;
-  /**
-   * @internal
-   */
-  public plural?: IXliffPlural;
-  /**
-   * @internal
-   */
-  public pluralTarget?: IXliffPlural;
-  /**
-   * @internal
-   */
 
+  /**
+   * @internal
+   */
   public formControl?: FormControl;
 
   constructor(
-    private formBuilder: FormBuilder,
+    private formBuilder: UntypedFormBuilder,
     @Inject(ControlContainer)
     public controlContainer: FormGroupDirective
   ) {}
@@ -67,21 +50,8 @@ export class TranslationPluralItemComponent implements OnChanges, OnDestroy {
       return;
     }
 
-    if (typeof this.source === 'string') {
-      this.text = this.source;
-    } else if (this.source.name === 'plural') {
-      this.plural = this.source;
-      this.pluralTarget = this.target as IXliffPlural;
-    } else {
-      this.interpolation = this.source;
-    }
-
-    if (!this.text) {
-      return;
-    }
-
-    // Display the target only if the target is a string (could be an object from a previous edition)
     const target = typeof this.target === 'string' ? this.target : '';
+
     if (this.id) {
       this.formControl = this.formBuilder.control(target, {
         nonNullable: true,